+++ Sources/Sentry/Public/SentryBreadcrumb.h
@@ -11,7 +11,7 @@ NS_SWIFT_NAME(Breadcrumb)
 /**
  * Level of breadcrumb
  */
-@property (nonatomic) enum SentryLevel level;
+@property (nonatomic) SentryLevel level;
 
 /**
  * Category of bookmark, can be any string
@@ -46,7 +46,7 @@ NS_SWIFT_NAME(Breadcrumb)
  * @param category String
  * @return SentryBreadcrumb
  */
-- (instancetype)initWithLevel:(enum SentryLevel)level category:(NSString *)category;
+- (instancetype)initWithLevel:(SentryLevel)level category:(NSString *)category;
 - (instancetype)init;
 + (instancetype)new NS_UNAVAILABLE;